Extract interface for deferred TaskSpec work
