+++ async/src/main/java/org/odk/collect/async/CoroutineAndWorkManagerScheduler.kt
@@ -1,18 +1,8 @@
 package org.odk.collect.async
 
-import android.net.NetworkCapabilities
-import android.net.NetworkRequest
-import androidx.work.Constraints
-import androidx.work.Data
-import androidx.work.ExistingPeriodicWorkPolicy
-import androidx.work.ExistingWorkPolicy
-import androidx.work.NetworkType
-import androidx.work.OneTimeWorkRequest
-import androidx.work.PeriodicWorkRequest
 import androidx.work.WorkInfo
 import androidx.work.WorkManager
 import kotlinx.coroutines.Dispatchers
-import java.util.concurrent.TimeUnit
 import kotlin.coroutines.CoroutineContext
 
 class CoroutineAndWorkManagerScheduler(
@@ -28,6 +18,7 @@ class CoroutineAndWorkManagerScheduler(
     ) // Needed for Java construction
 
     private val taskSpecRunner = WorkManagerTaskSpecImmediateRunner(workManager)
+    private val taskSpecScheduler = WorkManagerTaskSpecScheduler(workManager)
 
     override fun immediate(
         tag: String,
@@ -44,43 +35,7 @@ class CoroutineAndWorkManagerScheduler(
         inputData: Map<String, String>,
         networkConstraint: Scheduler.NetworkType?
     ) {
-        val networkRequest = NetworkRequest.Builder().apply {
-            addCapability(NetworkCapabilities.NET_CAPABILITY_INTERNET)
-
-            when (networkConstraint) {
-                Scheduler.NetworkType.WIFI -> addTransportType(NetworkCapabilities.TRANSPORT_WIFI)
-                Scheduler.NetworkType.CELLULAR -> addTransportType(NetworkCapabilities.TRANSPORT_CELLULAR)
-                else -> Unit
-            }
-        }.build()
-
-        val networkType = when (networkConstraint) {
-            Scheduler.NetworkType.WIFI -> NetworkType.UNMETERED
-            Scheduler.NetworkType.CELLULAR -> NetworkType.METERED
-            else -> NetworkType.CONNECTED
-        }
-
-        val constraints = Constraints
-            .Builder()
-            .setRequiredNetworkRequest(networkRequest, networkType)
-            .build()
-
-        val workManagerInputData = Data.Builder()
-            .putString(TaskSpecWorker.DATA_TASK_SPEC_CLASS, spec.javaClass.name)
-            .putBoolean(
-                TaskSpecWorker.DATA_CELLULAR_ONLY,
-                networkConstraint == Scheduler.NetworkType.CELLULAR
-            )
-            .putAll(inputData)
-            .build()
-
-        val workRequest = OneTimeWorkRequest.Builder(TaskSpecWorker::class.java)
-            .addTag(tag)
-            .setConstraints(constraints)
-            .setInputData(workManagerInputData)
-            .build()
-
-        workManager.beginUniqueWork(tag, ExistingWorkPolicy.REPLACE, workRequest).enqueue()
+        taskSpecScheduler.schedule(tag, spec, inputData, networkConstraint = networkConstraint)
     }
 
     override fun networkDeferredRepeat(
@@ -89,35 +44,7 @@ class CoroutineAndWorkManagerScheduler(
         repeatPeriod: Long,
         inputData: Map<String, String>
     ) {
-        val constraints = Constraints.Builder()
-            .setRequiredNetworkType(NetworkType.CONNECTED)
-            .build()
-
-        val workManagerInputData = Data.Builder()
-            .putString(TaskSpecWorker.DATA_TASK_SPEC_CLASS, spec.javaClass.name)
-            .putAll(inputData)
-            .build()
-
-        val builder = PeriodicWorkRequest.Builder(
-            TaskSpecWorker::class.java,
-            repeatPeriod,
-            TimeUnit.MILLISECONDS
-        )
-            .addTag(tag)
-            .setInputData(workManagerInputData)
-            .setConstraints(constraints)
-
-        spec.backoffPolicy?.let { backoffPolicy ->
-            spec.backoffDelay?.let { backoffDelay ->
-                builder.setBackoffCriteria(backoffPolicy, backoffDelay, TimeUnit.MILLISECONDS)
-            }
-        }
-
-        workManager.enqueueUniquePeriodicWork(
-            tag,
-            ExistingPeriodicWorkPolicy.REPLACE,
-            builder.build()
-        )
+        taskSpecScheduler.schedule(tag, spec, inputData, repeatPeriod = repeatPeriod)
     }
 
     override fun cancelDeferred(tag: String) {
