+++ async/src/main/java/org/odk/collect/async/TaskSpecScheduler.kt
@@ -1,0 +1,11 @@
+package org.odk.collect.async
+
+interface TaskSpecScheduler {
+    fun schedule(
+        tag: String,
+        spec: TaskSpec,
+        inputData: Map<String, String>,
+        networkConstraint: Scheduler.NetworkType? = null,
+        repeatPeriod: Long? = null
+    )
+}
